Make parsing literals possible
Literals/"unnamed types" aren't included in node-types.json in the same
way that named types are. Because of that lack of information it isn't
possible to generate strongly typed deserializers for traversing _all_
nodes in children. `NodeOrAny` is the solution.

Also in this is using a GAT to specify in the type system the type of
container that the child will be parsed into.